Finishing Draft Buddy

+++ src/main/java/com/ffa/controllers/UserController.java
@@ -39,7 +39,7 @@ public class UserController {
     //-------------------Retrieve Single User--------------------------------------------------------
       
     @RequestMapping(value = "/user/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<User> getUser(@PathVariable("id") int id) {
+    public ResponseEntity<User> getUser(@PathVariable("id") long id) {
         System.out.println("Fetching User with id " + id);
         User user = userService.findById(id);
         if (user == null) {
@@ -74,7 +74,7 @@ public class UserController {
     //------------------- Update a User --------------------------------------------------------
       
     @RequestMapping(value = "/user/{id}", method = RequestMethod.PUT)
-    public ResponseEntity<User> updateUser(@PathVariable("id") int id, @RequestBody User user) {
+    public ResponseEntity<User> updateUser(@PathVariable("id") long id, @RequestBody User user) {
         System.out.println("Updating User " + id);
           
         User currentUser = userService.findById(id);
@@ -95,7 +95,7 @@ public class UserController {
     //------------------- Delete a User --------------------------------------------------------
       
     @RequestMapping(value = "/user/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<User> deleteUser(@PathVariable("id") int id) {
+    public ResponseEntity<User> deleteUser(@PathVariable("id") long id) {
         System.out.println("Fetching & Deleting User with id " + id);
   
         User user = userService.findById(id);
