+++ src/main/java/com/ffa/controllers/UserService.java
@@ -4,11 +4,11 @@ import com.ffa.models.User;
  
 public interface UserService {
      
-    User findById(int id);
+    User findById(long id);
     User findByEmail(String email);
     void saveUser(User user);
     void updateUser(User user);
-    void deleteUserById(int id);
+    void deleteUserById(long id);
     List<User> findAllUsers(); 
     void deleteAllUsers();
     public boolean isUserExist(User user);
